+++ lib/src/annotations/annotations_reader.dart
@@ -25,46 +25,12 @@ class AnnotationReader<T> {
     return value;
   }
 
-  bool? getBoolean(String fieldName) {
-    bool? value;
-    for (var a in _annotations) {
-      final v = a.getField(fieldName)?.toBoolValue();
-      if (v != null) {
-        value = v;
-      }
-    }
-    return value;
-  }
-
   bool isSet(String fieldName) {
     for (var a in _annotations) {
       final v = a.getField(fieldName)?.toBoolValue();
       if (v != null && v) return true;
     }
     return false;
-  }
-
-  DartType? getType(String fieldName) {
-    DartType? value;
-    for (var a in _annotations) {
-      final v = a.getField(fieldName)?.toTypeValue();
-      if (v != null) {
-        value = v;
-      }
-    }
-    return value;
-  }
-
-  List<DartObject>? getList(String fieldName) {
-    List<DartObject>? list;
-    for (var a in _annotations) {
-      var v = a.getField(fieldName);
-      if (v != null) {
-        list ??= <DartObject>[];
-        list.addAll(v.toListValue()?.whereNotEmpty() ?? [v]);
-      }
-    }
-    return list;
   }
 
   static bool _isSquadronMarshaller(InterfaceType clazz) {